requirements/dev-requirements.txt
django-debug-toolbar
django-devserver>=0.7.0
coverage
guppy
werkzeug
line_profiler==1.0b3
git+git://github.com/dimagi/couchdbkit-debugpanel.git
Fabric==1.10.0
Sphinx==1.2b1
Jinja2==2.7.1
docutils==0.11
-e git+git://github.com/dimagi/luna.git@1.0.0#egg=luna
fixture
freezegun
